[Engine] - c++ script - IdMap 도입으로 컴포넌트 배열에 빈 컴포넌트가 없도록 변경 - 네이밍 및 코드 정리

+++ Engine/Source/RHI/DX11/CRD11Mesh.cpp
@@ -22,7 +22,7 @@ void CRD11Mesh::InitializePrimitive( const CRName& Name, const CRPrimitiveAsset&
     CRVertex::LoadFromPrimitiveData( PrimitiveData, vertice );
     
     VertexBuffer = GD11RM.GetVertexBuffer( Name );
-    VertexBuffer.lock()->Create( D3D11_USAGE_DEFAULT, 0, vertice.data(), sizeof( CRVertex ), vertice.size() );
+    VertexBuffer.lock()->Create( D3D11_USAGE_DEFAULT, 0, vertice.data(), sizeof( CRVertex ), (u32)vertice.size() );
 }
 
 //---------------------------------------------------------------------------------------------------------------------
